fix(logging): Add new strategy to initialize google cloud logger
