+++ src/utils/logging.ts
@@ -1,15 +1,42 @@
-import { Log } from '@google-cloud/logging'
+import { Log, Logging } from '@google-cloud/logging'
 import Container from 'typedi'
+import path from 'path'
+import fs from 'fs'
+import os from 'os'
 
 const metadata = {
   resource: { type: 'global' },
 }
 
+let loggerInstance: Log | undefined = undefined
+
+const getLogger = (): Log | undefined => {
+  if (!process.env.GOOGLE_APPLICATION_CREDENTIALS_KEY) {
+    console.error(
+      'Environment variable GOOGLE_APPLICATION_CREDENTIALS_KEY is not set'
+    )
+    return undefined
+  }
+
+  if (loggerInstance === undefined) {
+    console.log('Initializing logger')
+    const keyFile = path.join(os.tmpdir(), 'logging-sa-key.json')
+    fs.writeFileSync(
+      keyFile,
+      process.env.GOOGLE_APPLICATION_CREDENTIALS_KEY ?? ''
+    )
+    const gcpLoggingClient = new Logging({ keyFile })
+    loggerInstance = gcpLoggingClient.log('hosted-sessions')
+  }
+
+  return loggerInstance
+}
+
 export function log(params: {}, severity: string, error: string | {}) {
-  if (!Container.has('gcpLogger')) {
+  const logger = getLogger()
+  if (logger === undefined) {
     console.log('GCP log entry:', JSON.stringify({ params, severity, error }))
   } else {
-    const logger = Container.get<Log>('gcpLogger')
     logger
       .write(
         logger.entry({ ...metadata, severity: severity }, { params, error })
